pyproject.toml
[tool.poetry]
name = "fettler"
version = "0.1.0"
description = "Auto refresh cache of redis with MySQL binlog"
authors = ["long2ice <long2ice@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/long2ice/fettler"
repository = "https://github.com/long2ice/fettler.git"
documentation = "https://github.com/long2ice/fettler"
keywords = ["redis", "cache", "mysql", "binlog"]
packages = [
    { include = "fettler" }
]
include = ["CHANGELOG.md", "LICENSE", "README.md"]

[tool.poetry.dependencies]
python = "^3.7"
click = "*"
dynaconf = "*"
fastapi = "*"
uvicorn = "*"
aioredis = "*"
asyncmy = { git = "https://github.com/long2ice/asyncmy", branch = "dev" }
ciso8601 = "*"
loguru = "*"

[tool.poetry.dev-dependencies]
flake8 = "*"
isort = "*"
black = "^20.8b1"
pytest = "*"
pytest-xdist = "*"
pytest-asyncio = "*"
bandit = "*"
pytest-mock = "*"
httpx = "*"
tortoise-orm = { git = "https://github.com/tortoise/tortoise-orm", branch = "develop" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fettler = "fettler.cli:main"
